requirements.txt
matplotlib==3.6.2
numpy==1.24.1
pydicom==2.3.1
PySide6==6.4.2
